Fixes for civilian promotions
